Removed A11yOptions out of namespace.
