+++ ts/Accessibility/FocusBorder.ts
@@ -19,6 +19,9 @@
  * */
 
 import type Accessibility from './Accessibility';
+import type {
+    AccessibilityKeyboardNavigationFocusBorderOptions
+} from './Options/A11yOptions';
 import type { DOMElementType } from '../Core/Renderer/DOMElementType';
 import type SVGAttributes from '../Core/Renderer/SVG/SVGAttributes';
 
@@ -156,7 +159,7 @@ namespace FocusBorderComposition {
     ): void {
         const focusElement = this.focusElement,
             focusBorderOptions: (
-                Highcharts.AccessibilityKeyboardNavigationFocusBorderOptions
+                AccessibilityKeyboardNavigationFocusBorderOptions
             ) = this.options.accessibility.keyboardNavigation.focusBorder;
 
         if (focusElement) {
@@ -192,7 +195,7 @@ namespace FocusBorderComposition {
         focusElement?: DOMElementType
     ): void {
         const focusBorderOptions: (
-                Highcharts.AccessibilityKeyboardNavigationFocusBorderOptions
+                AccessibilityKeyboardNavigationFocusBorderOptions
             ) = this.options.accessibility.keyboardNavigation.focusBorder,
             browserFocusElement = focusElement || svgElement.element;
