+++ ts/Accessibility/Options/DeprecatedOptions.ts
@@ -69,7 +69,10 @@
  *
  * */
 
-import type A11yOptions from '../Options/Options';
+import type {
+    AxisAccessibilityOptions,
+    SeriesAccessibilityOptions
+} from './A11yOptions';
 import type Options from '../../Core/Options';
 import type Series from '../../Core/Series/Series';
 
@@ -90,18 +93,22 @@ const {
 declare module '../../Core/Axis/AxisOptions' {
     interface AxisOptions {
         /** @deprecated */
-        description?: A11yOptions.AxisAccessibilityOptions['description'];
+        description?: AxisAccessibilityOptions['description'];
     }
 }
 
 declare module '../../Core/Options'{
     interface Options {
         /** @deprecated */
-        exposeElementToA11y?: (
-            Highcharts.SeriesAccessibilityOptions['exposeAsGroupOnly']
-        );
+        exposeElementToA11y?: SeriesAccessibilityOptions['exposeAsGroupOnly'];
     }
 }
+
+/* *
+ *
+ *  Functions
+ *
+ * */
 
 /* eslint-disable valid-jsdoc */
 
@@ -369,5 +376,11 @@ function copyDeprecatedOptions(chart: Chart): void {
     copyDeprecatedKeyboardNavigationOptions(chart);
     copyDeprecatedLangOptions(chart);
 }
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
 
 export default copyDeprecatedOptions;
